requirements.txt
Django
djangorestframework
coverage
djangorestframework-simplejwt
django-cors-headers